create scipts to decode block & tx by id
